Modulators and demodulators no longer need to handle resampling or mixing
themselves---these operations are now the responsibility of the packet
modulator/demodulator.

This merge also contains a notable change to liquid: the demodulator callback is
now called when the header is received but before the packet has been
demodulated to see if we want to bother demodulating the packet. This should
reduce computational load when there is a lot of traffic on the air that a node
doesn't care about.

+++ src/phy/MultiOFDM.hh
@@ -90,14 +90,14 @@ public:
     virtual ~MultiOFDM() = default;
 
     // MultiChannel TX/RX requires oversampling by a factor of 2
-    double getMinRXRateOversample(void) const override
+    unsigned getMinRXRateOversample(void) const override
     {
-        return 2.0;
+        return 2;
     }
 
-    double getMinTXRateOversample(void) const override
+    unsigned getMinTXRateOversample(void) const override
     {
-        return 2.0;
+        return 2;
     }
 
 protected:
